Automatically run Pagerank on load
This commit modifies the `sourcecred load` command so that it
automatically runs the Pagerank command once the data has been loaded,
as was described in #1114. This will allow us to rewrite the frontend to
load the already-evaluated Pagerank results from the backend, in
addition to being convenient for writing other CLI commands that depend
on Pagerank results being available.

Test plan: The unit tests have been updated to check that calling `load`
invokes the `pagerank` command properly. Additionally, I ran the
following test as an experiment:

```
$ rm -rf $SOURCECRED_DIRECTORY;
$ yarn backend;
$ node bin/sourcecred.js load sourcecred/research;
$ ls $SOURCECRED_DIRECTORY/data/sourcecred/research
```

As expected, the `pagerankGraph.json` file was present in the directory;
additionally, the load command printed that it was running Pagerank and
that the task completed successfully.

Further progress on #967.